+++ src/main/java/blackjack/BlackJackApplication.java
@@ -1,6 +1,5 @@
 package blackjack;
 
-import blackjack.controller.BlackJackController;
 import blackjack.domain.gamer.Player;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
@@ -11,7 +10,7 @@ public class BlackJackApplication {
 
     public static void main(String[] args) {
         List<String> names = InputView.getNames();
-        BlackJackController controller = new BlackJackController(names);
+        BlackJackGame controller = new BlackJackGame(names);
 
         OutputView.printFirstCards(controller.getDealer(), controller.getPlayers());
 
@@ -22,18 +21,18 @@ public class BlackJackApplication {
         OutputView.printFinalResult(controller.createResult());
     }
 
-    private static void printAddtionalDrawDealr(BlackJackController controller) {
+    private static void printAddtionalDrawDealr(BlackJackGame controller) {
         controller.distributeAdditionalToDealer();
         OutputView.printAdditionalDrawDealer(controller.getDealerCardSize());
     }
 
-    private static void drawAdditionalCard(List<String> names, BlackJackController controller) {
+    private static void drawAdditionalCard(List<String> names, BlackJackGame controller) {
         for (String name : names) {
             drawCardToPlayer(controller, name);
         }
     }
 
-    private static void drawCardToPlayer(BlackJackController controller, String name) {
+    private static void drawCardToPlayer(BlackJackGame controller, String name) {
         while (controller.isDrawPossible(name, InputView::getAnswerOfAdditionalDraw)) {
             controller.distributeCardToPlayer(name);
             Player playerDtoByName = controller.findPlayerByName(name);
